Added fixes from pr. removed address variables. Refactored data source group code location
