+++ pkg/artifactory/datasource/group/datasource_artifactory_group.go
@@ -1,4 +1,4 @@
-package datasource
+package group
 
 import (
 	"golang.org/x/net/context"
