+++ pkg/artifactory/datasource/group/datasource_artifactory_group_test.go
@@ -1,4 +1,4 @@
-package datasource_test
+package group_test
 
 import (
 	"log"
